metrics/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
beautifulsoup4==4.13.3
blinker==1.8.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
decorator==4.4.2
deepface==0.0.93
facenet-pytorch==2.6.0
filelock==3.16.1
fire==0.7.0
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==25.2.10
fsspec==2025.3.0
gast==0.4.0
gdown==5.2.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.70.0
gunicorn==23.0.0
h5py==3.11.0
huggingface-hub==0.29.3
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
keras==2.13.1
libclang==18.1.1
lightning-utilities==0.11.9
lpips==0.1.4
Markdown==3.7
MarkupSafe==2.1.5
moviepy==1.0.3
mpmath==1.3.0
mtcnn==0.1.1
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pandas==2.0.3
pillow==10.2.0
proglog==0.1.10
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
retina-face==0.0.17
rsa==4.9
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
soupsieve==2.6
sympy==1.13.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.2.2
torch-fidelity==0.3.0
torchmetrics==1.5.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.46.3
triton==2.2.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
Werkzeug==3.0.6
wrapt==1.17.2
zipp==3.20.2
